backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.9.0.post0
pytz==2023.3
